fix: every fmt, clippy, and compile failure CI found in Phase 1
Nothing here changes behavior; two of these are the first honest corrections a
workspace that had never been compiled needed.

Formatting (cargo fmt, run 35136693454): app.rs import order, three over-long
expressions wrapped (set_rate if/else, the monitor chain, matches!+tooltip), the
MenuState literal expanded; model.rs sorted its Civil import and split the
over-100-char on_plate boolean; main.rs wrapped the settings-unreadable eprintln.

Compile errors (E0308, both Windows targets, identical):
- clock/format.rs:35 — Civil.month is u32, the Hinnant arithmetic yields i64; the
  sibling field day already cast, month now does too (m is provably 1..=12).
- placement.rs:183 — pick_monitor mixed Option<&Monitor> (via copied()) with
  Option<&&Monitor> (from first()); first() is copied() now, which is what the
  trailing .copied() implies the chain was written expecting.

Lints (many_single_char_names, -D warnings): relax() and advance() carried 5 and 8
single-char math locals respectively; renamed to pa/pb/delta/ratio/moved and
dt/gravity/cur/prev/brake/vel/moved/shrink, every use updated. No #[allow] anywhere.

Two rustfmt hunks (model.rs 18 and 195 regions) were truncated from the diff by the
old capture; if they persist, the chunked annotations will now show them in full.

+++ crates/hanglock-core/src/clock/format.rs
@@ -32,7 +32,7 @@ pub fn civil_from_unix_ms(ms: i64) -> Civil {
     let m = if mp < 10 { mp + 3 } else { mp - 9 };
     Civil {
         year: y + i64::from(if m <= 2 { 1 } else { 0 }),
-        month: m,
+        month: m as u32,
         day: d as u32,
         hour: (sod / 3600) as u32,
         minute: (sod / 60 % 60) as u32,
